NXDRIVE-1272: Move translator.py to nxdrive/, add short_name()
