+++ nxdrive/gui/conflicts.py
@@ -5,8 +5,8 @@ from typing import Any, Dict, List, Optional
 from PyQt5.QtCore import QSize, QUrl, pyqtSlot
 
 from .dialog import QMLDriveApi
-from .translator import Translator
 from .view import FileModel, NuxeoView
+from ..translator import Translator
 from ..utils import find_resource
 
 __all__ = ("ConflictsView",)
